requirements.txt
pandas
numpy 
seaborn
matplotlib
scikit-learn
catboost
xgboost
dill
flask
mlflow
psycopg2-binary
python-dotenv
SQLAlchemy
fastapi
uvicorn
pymysql
passlib[bcrypt]
python-jose
# -e .
